requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.83.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cohere==5.20.7
colorlog==6.10.1
cryptography==46.0.5
datapizza-ai==0.0.9
datapizza-ai-clients-anthropic==0.0.4
datapizza-ai-clients-google==0.0.7
datapizza-ai-clients-openai==0.0.9
datapizza-ai-core==0.0.22
datapizza-ai-embedders-cohere==0.0.3
datapizza-ai-embedders-openai==0.0.5
datapizza-ai-parsers-docling==0.0.8
datapizza-ai-tools-duckduckgo==0.0.4
datapizza-ai-vectorstores-qdrant==0.0.7
ddgs==9.10.0
defusedxml==0.7.1
dill==0.4.1
distro==1.9.0
docling==2.75.0
docling-core==2.66.0
docling-ibm-models==3.11.0
docling-parse==5.4.0
docstring-parser==0.17.0
dotenv==0.9.9
easyocr==1.7.2
et-xmlfile==2.0.0
fake-useragent==2.2.0
faker==40.5.1
fastavro==1.12.1
filelock==3.24.3
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.2.0
google-auth==2.48.0
google-genai==1.64.0
grpcio==1.78.1
h11==0.16.0
h2==4.3.0
hf-xet==1.3.1
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.2
hyperframe==6.1.0
idna==3.11
imageio==2.37.2
importlib-metadata==8.7.1
jinja2==3.1.6
jiter==0.13.0
jsonlines==4.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
latex2mathml==3.78.1
lazy-loader==0.4
lxml==6.0.2
markdown-it-py==4.0.0
marko==2.2.2
markupsafe==3.0.3
mcp==1.26.0
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
networkx==3.6.1
ninja==1.13.0
numpy==2.4.2
ocrmac==1.0.1
omegaconf==2.3.0
openai==2.22.0
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==26.0
pandas==2.3.3
pillow==12.1.1
pluggy==1.6.0
polyfactory==3.3.0
portalocker==3.2.0
primp==1.0.0
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pyasn1==0.6.2
pyasn1-modules==0.4.2
pyclipper==1.4.0
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.13.1
pygments==2.19.2
pyjwt==2.11.0
pylatexenc==2.10
pyobjc-core==12.1
pyobjc-framework-cocoa==12.1
pyobjc-framework-coreml==12.1
pyobjc-framework-quartz==12.1
pyobjc-framework-vision==12.1
pypdfium2==5.5.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.22
python-pptx==1.0.2
pytz==2025.2
pyyaml==6.0.3
qdrant-client==1.17.0
rapidocr==3.6.0
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
rtree==1.4.1
safetensors==0.7.0
scikit-image==0.26.0
scipy==1.17.1
semchunk==2.2.2
setuptools==82.0.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8.3
sse-starlette==3.2.0
starlette==0.52.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.4
tifffile==2026.2.24
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==4.57.6
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
typer==0.21.2
types-requests==2.32.4.20260107
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
websockets==15.0.1
xlsxwriter==3.2.9
yarl==1.22.0
zipp==3.23.0
